Unify the usage of round robin mechanism accross the codebase
